refactor: move client components to separate files and improve supabase integration
