+++ .gitignore
@@ -1,5 +1,5 @@
 # Dependencies
-/node_modules
+/node_modules/*
 /.pnp
 .pnp.js
 .pnpm-debug.log*
