+++ lib/supabase/client.ts
@@ -2,43 +2,37 @@
 
 import { createBrowserClient } from '@supabase/ssr'
 import type { Database } from '../types/database'
-import { type CookieOptions } from '@supabase/ssr'
+import type { CookieOptions } from '@supabase/ssr'
 
 export type { CookieOptions }
 
-const cookieStore = {
-  get: (name: string) => {
-    return document.cookie
-      .split('; ')
-      .find((row) => row.startsWith(`${name}=`))
-      ?.split('=')[1]
-  },
-  set: (name: string, value: string, options: CookieOptions) => {
-    let cookie = `${name}=${value}`
-    if (options.path) cookie += `; path=${options.path}`
-    if (options.maxAge) cookie += `; max-age=${options.maxAge}`
-    if (options.domain) cookie += `; domain=${options.domain}`
-    if (options.secure) cookie += '; secure'
-    if (options.sameSite) cookie += `; samesite=${options.sameSite}`
-    document.cookie = cookie
-  },
-  remove: (name: string, options: CookieOptions) => {
-    document.cookie = `${name}=; expires=Thu, 01 Jan 1970 00:00:00 GMT${
-      options.path ? `; path=${options.path}` : ''
-    }`
-  },
-}
-
-export const createClient = () =>
-  createBrowserClient<Database>(
+export const createClient = () => {
+  return createBrowserClient<Database>(
     process.env.NEXT_PUBLIC_SUPABASE_URL!,
     process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!,
     {
-      auth: {
-        autoRefreshToken: true,
-        persistSession: true,
-        detectSessionInUrl: true
+      cookies: {
+        get(name: string) {
+          if (typeof document === 'undefined') return ''
+          return document.cookie
+            .split('; ')
+            .find((row) => row.startsWith(`${name}=`))
+            ?.split('=')[1]
+        },
+        set(name: string, value: string, options: CookieOptions) {
+          if (typeof document === 'undefined') return
+          let cookie = `${name}=${value}`
+          if (options.path) cookie += `; path=${options.path}`
+          if (options.maxAge) cookie += `; max-age=${options.maxAge}`
+          if (options.domain) cookie += `; domain=${options.domain}`
+          if (options.secure) cookie += '; secure'
+          document.cookie = cookie
+        },
+        remove(name: string, options: CookieOptions) {
+          if (typeof document === 'undefined') return
+          document.cookie = `${name}=; max-age=0${options.path ? `; path=${options.path}` : ''}`
+        },
       },
-      cookies: cookieStore,
     }
   )
+}
