+++ middleware.ts
@@ -2,7 +2,7 @@ import { createServerClient, type CookieOptions } from '@supabase/ssr'
 import { NextResponse, type NextRequest } from 'next/server'
 
 export async function middleware(request: NextRequest) {
-  let response = NextResponse.next({
+  const response = NextResponse.next({
     request: {
       headers: request.headers,
     },
@@ -17,16 +17,6 @@ export async function middleware(request: NextRequest) {
           return request.cookies.get(name)?.value
         },
         set(name: string, value: string, options: CookieOptions) {
-          request.cookies.set({
-            name,
-            value,
-            ...options,
-          })
-          response = NextResponse.next({
-            request: {
-              headers: request.headers,
-            },
-          })
           response.cookies.set({
             name,
             value,
@@ -34,15 +24,6 @@ export async function middleware(request: NextRequest) {
           })
         },
         remove(name: string, options: CookieOptions) {
-          request.cookies.delete({
-            name,
-            ...options,
-          })
-          response = NextResponse.next({
-            request: {
-              headers: request.headers,
-            },
-          })
           response.cookies.delete({
             name,
             ...options,
@@ -53,16 +34,15 @@ export async function middleware(request: NextRequest) {
   )
 
   // Check auth status
-  const { data: { session } } = await supabase.auth.getSession()
+  const {
+    data: { session },
+  } = await supabase.auth.getSession()
 
   // Protect all routes except auth and public ones
   const isAuthPage = request.nextUrl.pathname.startsWith('/auth')
-  const isPublicPage = [
-    '/_next',
-    '/images',
-    '/api/public',
-    '/favicon.ico'
-  ].some(path => request.nextUrl.pathname.startsWith(path))
+  const isPublicPage = ['/_next', '/images', '/api/public', '/favicon.ico'].some((path) =>
+    request.nextUrl.pathname.startsWith(path)
+  )
 
   if (!session && !isAuthPage && !isPublicPage) {
     return NextResponse.redirect(new URL('/auth', request.url))
@@ -73,17 +53,4 @@ export async function middleware(request: NextRequest) {
   }
 
   return response
-}
-
-export const config = {
-  matcher: [
-    /*
-     * Match all request paths except for the ones starting with:
-     * - _next/static (static files)
-     * - _next/image (image optimization files)
-     * - favicon.ico (favicon file)
-     * Feel free to modify this pattern to include more paths.
-     */
-    '/((?!_next/static|_next/image|favicon.ico).*)',
-  ],
 }
