+++ package.json
@@ -4,7 +4,7 @@
   "private": true,
   "type": "module",
   "engines": {
-    "node": "22.x",
+    "node": ">=20.0.0",
     "pnpm": ">=8.0.0"
   },
   "scripts": {
@@ -19,99 +19,91 @@
     "format": "prettier --write .",
     "format:check": "prettier --check .",
     "typecheck": "tsc --noEmit",
-    "ci": "pnpm run lint && pnpm run typecheck && pnpm run test:coverage && pnpm run format:check",
-    "clean": "rm -rf .next node_modules",
-    "postinstall": "husky install",
     "prepare": "husky install"
   },
   "dependencies": {
-    "@hookform/resolvers": "^3.3.4",
+    "@hookform/resolvers": "^3.10.0",
     "@radix-ui/react-avatar": "^1.1.2",
-    "@radix-ui/react-checkbox": "^1.0.4",
+    "@radix-ui/react-checkbox": "^1.1.3",
     "@radix-ui/react-dialog": "^1.1.4",
     "@radix-ui/react-dropdown-menu": "^2.1.4",
-    "@radix-ui/react-icons": "^1.3.0",
+    "@radix-ui/react-icons": "^1.3.2",
     "@radix-ui/react-label": "^2.1.1",
     "@radix-ui/react-popover": "^1.1.4",
-    "@radix-ui/react-progress": "1.1.1",
-    "@radix-ui/react-scroll-area": "1.2.2",
+    "@radix-ui/react-progress": "^1.1.1",
+    "@radix-ui/react-scroll-area": "^1.2.2",
     "@radix-ui/react-select": "^2.1.4",
+    "@radix-ui/react-separator": "^1.1.1",
     "@radix-ui/react-slot": "^1.1.1",
-    "@radix-ui/react-tabs": "^1.0.4",
+    "@radix-ui/react-tabs": "^1.1.2",
     "@radix-ui/react-toast": "^1.2.4",
     "@supabase/auth-ui-react": "^0.4.7",
     "@supabase/auth-ui-shared": "^0.1.8",
-    "@supabase/ssr": "^0.1.0",
-    "@supabase/supabase-js": "^2.47.12",
-    "@tanstack/react-table": "8.20.6",
-    "@tanstack/table-core": "8.20.5",
-    "@types/react-pdf": "7.0.0",
-    "chart.js": "^4.4.1",
+    "@supabase/ssr": "^0.5.2",
+    "@supabase/supabase-js": "^2.47.15",
+    "@tanstack/react-table": "^8.20.6",
+    "chart.js": "^4.4.7",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
-    "cmdk": "^1.0.4",
     "date-fns": "^3.6.0",
-    "iterall": "^1.3.0",
     "lucide-react": "^0.471.1",
-    "next": "15.1.4",
+    "next": "14.2.23",
     "next-themes": "^0.4.4",
-    "react": "19.0.0",
-    "react-chartjs-2": "^5.2.0",
+    "react": "^18.3.1",
+    "react-chartjs-2": "^5.3.0",
     "react-day-picker": "^8.10.1",
-    "react-dom": "19.0.0",
-    "react-dropzone": "14.3.5",
-    "react-hook-form": "^7.49.3",
-    "react-pdf": "9.2.1",
-    "recharts": "2.15.0",
-    "resend": "4.0.1",
+    "react-dom": "^18.3.1",
+    "react-dropzone": "^14.3.5",
+    "react-hook-form": "^7.54.2",
+    "react-pdf": "^7.7.3",
+    "recharts": "^2.15.0",
+    "resend": "^2.1.0",
     "tailwind-merge": "^2.6.0",
     "tailwindcss-animate": "^1.0.7",
-    "xlsx": "0.18.5",
-    "zod": "^3.22.4"
+    "xlsx": "^0.18.5",
+    "zod": "^3.24.1"
   },
   "devDependencies": {
     "@playwright/test": "1.49.1",
-    "@storybook/addon-essentials": "8.5.0",
-    "@storybook/builder-webpack5": "8.5.0",
-    "@storybook/manager-webpack5": "6.5.16",
-    "@storybook/react": "8.5.0",
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^14.3.1",
-    "@testing-library/user-event": "^14.5.2",
-    "@types/d3": "7.4.3",
-    "@types/jest": "29.5.14",
-    "@types/node": "^20.17.11",
-    "@types/react": "19.0.6",
-    "@types/react-dom": "19.0.3",
+    "@testing-library/user-event": "^14.6.0",
+    "@types/d3": "^7.4.3",
+    "@types/node": "^20.17.14",
+    "@types/react": "^18.3.18",
+    "@types/react-dom": "^18.3.5",
+    "@types/testing-library__jest-dom": "^5.14.9",
+    "@types/xlsx": "^0.0.36",
     "@typescript-eslint/eslint-plugin": "^6.21.0",
     "@typescript-eslint/parser": "^6.21.0",
     "@vitejs/plugin-react": "^4.3.4",
-    "@vitest/coverage-v8": "^2.1.8",
-    "@vitest/ui": "^2.1.8",
-    "autoprefixer": "^10.4.16",
-    "eslint": "^8.56.0",
-    "eslint-config-next": "^15.1.4",
+    "@vitest/coverage-v8": "^1.6.0",
+    "@vitest/ui": "^1.6.0",
+    "autoprefixer": "^10.4.20",
+    "eslint": "^8.57.1",
+    "eslint-config-next": "14.2.23",
     "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-jest-dom": "^5.1.0",
-    "eslint-plugin-testing-library": "^6.2.0",
-    "husky": "^9.0.6",
+    "eslint-plugin-jest-dom": "^5.5.0",
+    "eslint-plugin-react": "^7.37.4",
+    "eslint-plugin-react-hooks": "^4.6.2",
+    "eslint-plugin-testing-library": "^6.5.0",
+    "husky": "^8.0.3",
     "jsdom": "^23.2.0",
-    "lint-staged": "^15.3.0",
-    "postcss": "^8.4.33",
+    "lint-staged": "^15.4.0",
+    "postcss": "^8.5.1",
     "prettier": "^3.4.2",
-    "prettier-plugin-tailwindcss": "^0.5.11",
-    "supabase": "^2.2.1",
-    "tailwindcss": "^3.4.1",
+    "prettier-plugin-tailwindcss": "^0.6.10",
+    "tailwindcss": "^3.4.17",
     "typescript": "^5.7.3",
-    "vite-tsconfig-paths": "5.1.4",
-    "vitest": "^2.1.8"
+    "vite": "^5.4.11",
+    "vitest": "^1.6.0"
   },
   "lint-staged": {
     "*.{js,jsx,ts,tsx}": [
       "eslint --fix",
       "prettier --write"
     ],
-    "*.{json,md,yml}": [
+    "*.{json,css,md}": [
       "prettier --write"
     ]
   }
